requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
funcy==1.17
msgpack
neo4j==4.4.4
numpy
pandas
pika==1.2.1
# psycopg2==2.9.3
psycopg2_binary==2.9.3
PyYAML==6.0
redis==4.2.2
requests==2.26.0
SQLAlchemy==1.4.22
tqdm==4.62.3
# -i https://download.pytorch.org/whl/cu117
# torch
# torchvision
# torchaudio
cuda-python==11.7.1
# -f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cu115
torchvision==0.12.0+cu115
torchaudio==0.11.0+cu115
dgl==0.4.3
dglke==0.1.2